# server/router/auth-provider-router.go
package router

import (
	"log"
	"net/http"

	"github.com/gorilla/mux"

	. "github.com/seatsurfing/seatsurfing/server/api"
	. "github.com/seatsurfing/seatsurfing/server/repository"
	. "github.com/seatsurfing/seatsurfing/server/util"
)

type AuthProviderRouter struct {
}

type CreateAuthProviderRequest struct {
	Name                   string `json:"name" validate:"required,max=256"`
	ProviderType           int    `json:"providerType" validate:"required"`
	AuthURL                string `json:"authUrl" validate:"required,max=512"`
	TokenURL               string `json:"tokenUrl" validate:"required,max=512"`
	AuthStyle              int    `json:"authStyle"`
	Scopes                 string `json:"scopes" validate:"required,max=256"`
	UserInfoURL            string `json:"userInfoUrl" validate:"required,max=512"`
	UserInfoEmailField     string `json:"userInfoEmailField" validate:"required,max=256"`
	UserInfoFirstnameField string `json:"userInfoFirstnameField" validate:"max=256"`
	UserInfoLastnameField  string `json:"userInfoLastnameField" validate:"max=256"`
	UserInfoGroupsField    string `json:"userInfoGroupsField" validate:"max=256"`
	ClientID               string `json:"clientId" validate:"required,max=256"`
	ClientSecret           string `json:"clientSecret,omitempty" validate:"max=256"`
	LogoutURL              string `json:"logoutUrl" validate:"max=256"`
	ProfilePageURL         string `json:"profilePageUrl" validate:"max=256"`
}

type GetAuthProviderResponse struct {
	ID             string `json:"id"`
	ReadOnly       bool   `json:"readOnly"`
	OrganizationID string `json:"organizationId"`
	CreateAuthProviderRequest
}

type GetAuthProviderPublicResponse struct {
	ID   string `json:"id"`
	Name string `json:"name"`
}

func (router *AuthProviderRouter) SetupRoutes(s *mux.Router) {
	s.HandleFunc("/org/{id}", router.listPublicForOrg).Methods("GET")
	s.HandleFunc("/{id}/mapping", router.getMappings).Methods("GET")
	s.HandleFunc("/{id}/mapping", router.setMappings).Methods("PUT")
	s.HandleFunc("/{id}", router.getOne).Methods("GET")
	s.HandleFunc("/{id}", router.update).Methods("PUT")
	s.HandleFunc("/{id}", router.delete).Methods("DELETE")
	s.HandleFunc("/", router.create).Methods("POST")
	s.HandleFunc("/", router.getAll).Methods("GET")
}

func (router *AuthProviderRouter) listPublicForOrg(w http.ResponseWriter, r *http.Request) {
	vars := mux.Vars(r)
	org, err := GetOrganizationRepository().GetOne(vars["id"])
	if err != nil {
		log.Println(err)
		SendNotFound(w)
		return
	}
	list, err := GetAuthProviderRepository().GetAll(org.ID)
	if err != nil {
		log.Println(err)
		SendInternalServerError(w)
		return
	}
	res := []*GetAuthProviderPublicResponse{}
	for _, e := range list {
		m := &GetAuthProviderPublicResponse{}
		m.ID = e.ID
		m.Name = e.Name
		res = append(res, m)
	}
	SendJSON(w, res)
}

func (router *AuthProviderRouter) getOne(w http.ResponseWriter, r *http.Request) {
	vars := mux.Vars(r)
	e, err := GetAuthProviderRepository().GetOne(vars["id"])
	if err != nil {
		log.Println(err)
		SendNotFound(w)
		return
	}
	user := GetRequestUser(r)
	if !HasPermission(user, e.OrganizationID, PermissionAuthProviders, PermissionLevelAdmin) {
		SendForbidden(w)
		return
	}
	res := router.copyToRestModel(e)
	SendJSON(w, res)
}

func (router *AuthProviderRouter) getAll(w http.ResponseWriter, r *http.Request) {
	user := GetRequestUser(r)
	if !HasPermission(user, user.OrganizationID, PermissionAuthProviders, PermissionLevelAdmin) {
		SendForbidden(w)
		return
	}
	list, err := GetAuthProviderRepository().GetAll(user.OrganizationID)
	if err != nil {
		log.Println(err)
		SendInternalServerError(w)
		return
	}
	res := []*GetAuthProviderResponse{}
	for _, e := range list {
		m := router.copyToRestModel(e)
		res = append(res, m)
	}
	SendJSON(w, res)
}

func (router *AuthProviderRouter) validateCreateAuthProviderRequest(m *CreateAuthProviderRequest) bool {
	if m.ProviderType != int(OAuth2) {
		return false
	}
	if m.AuthStyle < 0 || m.AuthStyle > 2 {
		return false
	}
	if !ValidateURL(m.AuthURL) || !ValidateURL(m.TokenURL) || !ValidateURL(m.UserInfoURL) {
		return false
	}
	if m.LogoutURL != "" && !ValidateURL(m.LogoutURL) {
		return false
	}
	if m.ProfilePageURL != "" && !ValidateURL(m.ProfilePageURL) {
		return false
	}
	return true
}

func (router *AuthProviderRouter) update(w http.ResponseWriter, r *http.Request) {
	var m CreateAuthProviderRequest
	if UnmarshalValidateBody(r, &m) != nil {
		SendBadRequest(w)
		return
	}
	if !router.validateCreateAuthProviderRequest(&m) {
		SendBadRequest(w)
		return
	}
	vars := mux.Vars(r)
	e, err := GetAuthProviderRepository().GetOne(vars["id"])
	if err != nil {
		SendBadRequest(w)
		return
	}
	if e.ReadOnly {
		SendForbidden(w)
		return
	}
	user := GetRequestUser(r)
	if !HasPermission(user, e.OrganizationID, PermissionAuthProviders, PermissionLevelAdmin) {
		SendForbidden(w)
		return
	}

	existingAuthProvider, err := GetAuthProviderRepository().GetByName(e.OrganizationID, m.Name)
	if err == nil && existingAuthProvider != nil && existingAuthProvider.ID != e.ID {
		SendAlreadyExistsCode(w, ResponseCodeAuthProviderAlreadyExists)
		return
	}

	// keep existing client secret if no new secret was set or encrypt the new one
	if m.ClientSecret == "" {
		m.ClientSecret = e.ClientSecret
	} else {
		ClientSecretEncrypted, err := EncryptString(m.ClientSecret)
		if err != nil {
			log.Println("Error encrypting client secret")
		}
		m.ClientSecret = ClientSecretEncrypted
	}

	eNew := router.copyFromRestModel(&m)

	eNew.ID = e.ID
	eNew.OrganizationID = e.OrganizationID
	eNew.ReadOnly = e.ReadOnly
	if err := GetAuthProviderRepository().Update(eNew); err != nil {
		log.Println(err)
		SendInternalServerError(w)
		return
	}
	SendUpdated(w)
}

func (router *AuthProviderRouter) delete(w http.ResponseWriter, r *http.Request) {
	vars := mux.Vars(r)
	e, err := GetAuthProviderRepository().GetOne(vars["id"])
	if err != nil {
		log.Println(err)
		SendNotFound(w)
		return
	}
	user := GetRequestUser(r)
	if !HasPermission(user, e.OrganizationID, PermissionAuthProviders, PermissionLevelAdmin) {
		SendForbidden(w)
		return
	}
	if e.ReadOnly {
		SendForbidden(w)
		return
	}

	// Check if any users are bound to this auth provider
	hasUsers, err := GetUserRepository().HasAnyUserWithAuthProvider(e.ID)
	if err != nil {
		log.Println(err)
		SendInternalServerError(w)
		return
	}
	if hasUsers {
		SendBadRequest(w)
		return
	}

	if err := GetAuthProviderRepository().Delete(e); err != nil {
		log.Println(err)
		SendInternalServerError(w)
		return
	}
	SendUpdated(w)
}

func (router *AuthProviderRouter) create(w http.ResponseWriter, r *http.Request) {
	var m CreateAuthProviderRequest
	if err := UnmarshalValidateBody(r, &m); err != nil {
		log.Println(err)
		SendBadRequest(w)
		return
	}
	if !router.validateCreateAuthProviderRequest(&m) || m.ClientSecret == "" {
		SendBadRequest(w)
		return
	}

	user := GetRequestUser(r)
	if !HasPermission(user, user.OrganizationID, PermissionAuthProviders, PermissionLevelAdmin) {
		SendForbidden(w)
		return
	}

	// encrypt client secret
	ClientSecretEncrypted, err := EncryptString(m.ClientSecret)
	if err != nil {
		log.Println("Error encrypting client secret")
	}
	m.ClientSecret = ClientSecretEncrypted

	e := router.copyFromRestModel(&m)
	e.OrganizationID = user.OrganizationID
	e.ReadOnly = false

	featureAuthProviders, _ := GetSettingsRepository().GetBool(user.OrganizationID, SettingFeatureAuthProviders.Name)
	if !featureAuthProviders {
		SendPaymentRequired(w)
		return
	}

	existingAuthProvider, err := GetAuthProviderRepository().GetByName(e.OrganizationID, e.Name)
	if err == nil && existingAuthProvider != nil {
		SendAlreadyExistsCode(w, ResponseCodeAuthProviderAlreadyExists)
		return
	}

	if err := GetAuthProviderRepository().Create(e); err != nil {
		log.Println(err)
		SendInternalServerError(w)
		return
	}
	SendCreated(w, e.ID)
}

func (router *AuthProviderRouter) copyFromRestModel(m *CreateAuthProviderRequest) *AuthProvider {
	e := &AuthProvider{}
	e.Name = m.Name
	e.ClientID = m.ClientID
	e.ClientSecret = m.ClientSecret
	e.AuthURL = m.AuthURL
	e.TokenURL = m.TokenURL
	e.AuthStyle = m.AuthStyle
	e.Scopes = m.Scopes
	e.UserInfoURL = m.UserInfoURL
	e.UserInfoEmailField = m.UserInfoEmailField
	e.UserInfoFirstnameField = m.UserInfoFirstnameField
	e.UserInfoLastnameField = m.UserInfoLastnameField
	e.UserInfoGroupsField = m.UserInfoGroupsField
	e.ProviderType = m.ProviderType
	e.LogoutURL = m.LogoutURL
	e.ProfilePageURL = m.ProfilePageURL
	return e
}

func (router *AuthProviderRouter) copyToRestModel(e *AuthProvider) *GetAuthProviderResponse {
	m := &GetAuthProviderResponse{}
	m.ID = e.ID
	m.OrganizationID = e.OrganizationID
	m.Name = e.Name
	m.ClientID = e.ClientID
	e.ClientSecret = "" // client secret is not exposed to the client
	m.AuthURL = e.AuthURL
	m.TokenURL = e.TokenURL
	m.AuthStyle = e.AuthStyle
	m.Scopes = e.Scopes
	m.UserInfoURL = e.UserInfoURL
	m.UserInfoEmailField = e.UserInfoEmailField
	m.UserInfoFirstnameField = e.UserInfoFirstnameField
	m.UserInfoLastnameField = e.UserInfoLastnameField
	m.UserInfoGroupsField = e.UserInfoGroupsField
	m.ProviderType = e.ProviderType
	m.LogoutURL = e.LogoutURL
	m.ProfilePageURL = e.ProfilePageURL
	m.ReadOnly = e.ReadOnly
	return m
}

// ─── Identity provider mappings ──────────────────────────────────────────────

type AuthProviderMappingModel struct {
	ClaimValue string `json:"claimValue" validate:"required,max=512"`
	TargetType string `json:"targetType" validate:"required,oneof=role group"`
	TargetID   string `json:"targetId" validate:"required,uuid"`
}

type SetAuthProviderMappingsRequest struct {
	Mappings []*AuthProviderMappingModel `json:"mappings" validate:"dive"`
}

// getMappings returns the provider's claim-to-role and claim-to-group mappings.
func (router *AuthProviderRouter) getMappings(w http.ResponseWriter, r *http.Request) {
	e := router.loadProvider(w, r)
	if e == nil {
		return
	}
	list, err := GetAuthProviderMappingRepository().GetAll(e.ID)
	if err != nil {
		log.Println(err)
		SendInternalServerError(w)
		return
	}
	res := []*AuthProviderMappingModel{}
	for _, m := range list {
		res = append(res, &AuthProviderMappingModel{
			ClaimValue: m.ClaimValue,
			TargetType: m.TargetType,
			TargetID:   m.TargetID,
		})
	}
	SendJSON(w, res)
}

// setMappings replaces the provider's mappings. Each target must belong to the
// same organization as the provider, and a mapping may not grant a role the
// caller could not grant directly.
func (router *AuthProviderRouter) setMappings(w http.ResponseWriter, r *http.Request) {
	e := router.loadProvider(w, r)
	if e == nil {
		return
	}
	user := GetRequestUser(r)
	if !CheckPermission(w, user, e.OrganizationID, PermissionRoles, PermissionLevelAdmin) {
		return
	}
	var req SetAuthProviderMappingsRequest
	if UnmarshalValidateBody(r, &req) != nil {
		SendBadRequest(w)
		return
	}
	for _, item := range req.Mappings {
		switch item.TargetType {
		case AuthProviderMappingTargetRole:
			role, err := GetRoleRepository().GetOne(item.TargetID)
			if err != nil || role.OrganizationID != e.OrganizationID {
				SendBadRequest(w)
				return
			}
			if !CanGrantPermissions(user, e.OrganizationID, role.Permissions) {
				SendBadRequestCode(w, ResponseCodeRoleEscalationNotAllowed)
				return
			}
		case AuthProviderMappingTargetGroup:
			group, err := GetGroupRepository().GetOne(item.TargetID)
			if err != nil || group.OrganizationID != e.OrganizationID {
				SendBadRequest(w)
				return
			}
		default:
			SendBadRequest(w)
			return
		}
	}
	if err := GetAuthProviderMappingRepository().DeleteAllForProvider(e.ID); err != nil {
		log.Println(err)
		SendInternalServerError(w)
		return
	}
	for _, item := range req.Mappings {
		mapping := &AuthProviderMapping{
			AuthProviderID: e.ID,
			ClaimValue:     item.ClaimValue,
			TargetType:     item.TargetType,
			TargetID:       item.TargetID,
		}
		if err := GetAuthProviderMappingRepository().Create(mapping); err != nil {
			log.Println(err)
			SendInternalServerError(w)
			return
		}
	}
	SendUpdated(w)
}

// loadProvider resolves the provider named in the path, enforcing the
// auth_providers permission and organization scoping.
func (router *AuthProviderRouter) loadProvider(w http.ResponseWriter, r *http.Request) *AuthProvider {
	user := GetRequestUser(r)
	vars := mux.Vars(r)
	e, err := GetAuthProviderRepository().GetOne(vars["id"])
	if err != nil {
		SendNotFound(w)
		return nil
	}
	if !HasPermission(user, e.OrganizationID, PermissionAuthProviders, PermissionLevelAdmin) {
		SendForbidden(w)
		return nil
	}
	return e
}