# specs/openapi.yaml
openapi: 3.1.0
info:
  title: Seatsurfing REST API
  description: |
    Seatsurfing provides REST APIs used by the mobile and web app as well as the administrator web interface.

    ## Authorization checks

    Each API endpoint is protected by authorization checks:
    1. **Permission check**: the permissions granted by the roles assigned to the calling user
    2. **Organization check**: organization membership - no request reaches data belonging to another organization

    ## Permissions

    Access is granted by roles, which an organization's administrators define. A role grants each
    functionality at a level, and a user may hold several roles, holding the highest level any of
    them grants.

    Levels: `0` no access, `10` read only, `20` limited, `30` full access. Not every level is
    meaningful for every permission; `GET /role/permissions` reports the levels each one offers.

    Permission keys: `areas`, `space_attributes`, `bookings`, `approvals`, `analytics`,
    `presence_report`, `users`, `groups`, `roles`, `service_accounts`, `auth_providers`,
    `org_settings`, `audit_log`. Plugins contribute further keys prefixed with `plugin.`.

    `analytics` covers aggregate occupancy figures; `presence_report` covers who was present
    and when. They are separate because the latter is personal data, and each is additionally
    subject to an organization-wide setting (`hide_stats` and `hide_reports`) that withholds it
...

...

...

    ## Authentication

...

...

...

...

...
  license:
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained via login or refresh endpoints
    BasicAuth:
      type: http
      scheme: basic
      description: Service account authentication using `{orgId}_{email}` as username and the service account password
    ServiceAccountBearerAuth:
      type: http
      scheme: bearer
      description: Service account authentication using a long-lived API token generated via `POST /user/{id}/api-token`. The raw token is presented as a Bearer token. Read-only service accounts may only use this on `GET` requests.
  schemas:
...
    AuthPreflightResponse:
      type: object
      properties:
        organization:
          $ref: "#/components/schemas/GetOrganizationResponse"
        authProviders:
          type: array
          items:
            $ref: "#/components/schemas/GetAuthProviderPublicResponse"
        requirePassword:
          type: boolean
          description: Whether any user has a password set
        disablePasswordLogin:
          type: boolean
          description: Whether password login is disabled globally
        domain:
          type: string
          description: Primary domain of the organization
...
    GetOrganizationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        email:
          type: string
          format: email
        language:
          type: string
...
    CreateAuthProviderRequest:
      type: object
      required:
        [
          name,
          providerType,
          authUrl,
          tokenUrl,
          scopes,
          userInfoUrl,
          userInfoEmailField,
          clientId,
          clientSecret,
        ]
      properties:
        name:
          type: string
        providerType:
          type: integer
        authUrl:
          type: string
          format: uri
        tokenUrl:
          type: string
          format: uri
        authStyle:
          type: integer
        scopes:
          type: string
          description: Comma-separated list of OAuth scopes
        userInfoUrl:
          type: string
          format: uri
        userInfoEmailField:
          type: string
        userInfoFirstnameField:
          type: string
        userInfoGroupsField:
          type: string
          maxLength: 256
          description: Name of the claim listing the user's groups, used to assign roles and group memberships automatically. Empty disables it.
        userInfoLastnameField:
          type: string
        clientId:
          type: string
        clientSecret:
          type: string
        logoutUrl:
          type: string
          format: uri
          description: "Supports placeholder: {logoutRedirectUri}"
        profilePageUrl:
          type: string
          format: uri
    GetAuthProviderResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        readOnly:
          type: boolean
          description: Whether this provider was auto-configured and cannot be edited
        name:
          type: string
        providerType:
          type: integer
        authUrl:
          type: string
        tokenUrl:
          type: string
        authStyle:
          type: integer
        scopes:
          type: string
        userInfoUrl:
          type: string
        userInfoEmailField:
          type: string
        userInfoFirstnameField:
          type: string
        userInfoLastnameField:
          type: string
        clientId:
          type: string
        logoutUrl:
          type: string
        profilePageUrl:
          type: string
    GetAuthProviderPublicResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string

    # --- Auth Events ---
...
  responses:
    Created:
      description: Object created successfully
      headers:
        X-Object-ID:
          schema:
            type: string
          description: ID of the created object
    Updated:
      description: Object updated or deleted successfully (204 No Content)
    BadRequest:
      description: Malformed request or incomplete JSON
      headers:
        X-Error-Code:
          schema:
            type: integer
          description: Custom error code (see error codes table)
...
    Forbidden:
      description: Authorized, but user has no access to requested resource
    NotFound:
      description: Object not found
...
    PaymentRequired:
      description: Feature not available (license/plan restriction)
...
paths:
...
  /auth/org/{domain}:
    get:
      tags: [Authentication]
      summary: Get organization details by domain
      description: Returns organization details and available auth providers for a domain. Used for the login preflight flow.
      operationId: getOrgDetails
      security: []
      parameters:
        - name: domain
          in: path
          required: true
          schema:
            type: string
          description: Organization domain
      responses:
        "200":
          description: Organization preflight data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthPreflightResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /auth-provider/:
    get:
      tags: [Auth Providers]
      summary: Get all auth providers
      description: Returns all authentication providers for the organization. Requires `auth_providers` at full access.
      operationId: getAllAuthProviders
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of auth providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetAuthProviderResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Auth Providers]
      summary: Create an auth provider
      description: Creates a new OAuth/OIDC authentication provider. Requires `auth_providers` at full access. Requires the Auth Providers feature.
      operationId: createAuthProvider
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAuthProviderRequest"
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "402":
          $ref: "#/components/responses/PaymentRequired"
        "403":
          $ref: "#/components/responses/Forbidden"
  /auth-provider/{id}:
    get:
      tags: [Auth Providers]
      summary: Get an auth provider
      description: Returns a single authentication provider by ID. Requires `auth_providers` at full access.
      operationId: getAuthProvider
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Auth provider details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAuthProviderResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Auth Providers]
      summary: Update an auth provider
      description: Updates an authentication provider. Read-only providers cannot be updated. Requires `auth_providers` at full access.
      operationId: updateAuthProvider
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAuthProviderRequest"
      responses:
        "204":
          $ref: "#/components/responses/Updated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [Auth Providers]
      summary: Delete an auth provider
      description: Deletes an authentication provider. Read-only providers cannot be deleted. Cannot delete providers that have users bound to them. Requires `auth_providers` at full access.
      operationId: deleteAuthProvider
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          $ref: "#/components/responses/Updated"
        "400":
          description: Provider has users bound to it
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /auth-provider/org/{id}:
    get:
      tags: [Auth Providers]
      summary: Get public auth providers for organization
      description: Returns the public auth provider information (ID and name only) for an organization. No authentication required (whitelisted route).
      operationId: getPublicAuthProviders
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Organization ID
      responses:
        "200":
          description: List of public auth providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetAuthProviderPublicResponse"
        "404":
          $ref: "#/components/responses/NotFound"

  # ===========================
  # Auth Events
  # ===========================
...